Refactor UI and state handling.
