+++ app/src/main/java/se/koditoriet/snout/ui/components/LoadingOverlay.kt
@@ -1,22 +1,20 @@
-package se.koditoriet.snout.ui.screens
+package se.koditoriet.snout.ui.components
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.window.Dialog
-import se.koditoriet.snout.ui.components.LoadingSpinner
 
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun LoadingScreen(isLoadingState: MutableState<Boolean>) {
-    if (isLoadingState.value) {
+fun LoadingOverlay(isLoadingState: Boolean) {
+    if (isLoadingState) {
         Dialog(onDismissRequest = {}) {
             Column(
                 modifier = Modifier
