+++ app/src/main/java/se/koditoriet/snout/ui/screens/main/passkeys/sheets/PasskeyActionsSheet.kt
@@ -1,37 +1,37 @@
-package se.koditoriet.snout.ui.sheets
+package se.koditoriet.snout.ui.screens.main.passkeys.sheets
 
 import BottomSheetAction
 import BottomSheetContextualHeader
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.AccountBox
 import androidx.compose.material.icons.filled.DeleteForever
 import androidx.compose.material.icons.filled.Edit
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.runtime.Composable
 import se.koditoriet.snout.appStrings
-import se.koditoriet.snout.vault.TotpSecret
+import se.koditoriet.snout.ui.components.PasskeyIcon
+import se.koditoriet.snout.vault.Passkey
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun SecretActionsSheet(
-    totpSecret: TotpSecret,
-    onEditMetadata: (TotpSecret) -> Unit,
-    onDeleteSecret: (TotpSecret) -> Unit,
+fun PasskeyActionsSheet(
+    passkey: Passkey,
+    onEditDisplayName: (Passkey) -> Unit,
+    onDeletePasskey: (Passkey) -> Unit,
 ) {
-    val screenStrings = appStrings.secretsScreen
+    val screenStrings = appStrings.managePasskeysScreen
     BottomSheetContextualHeader(
-        heading = totpSecret.issuer,
-        subheading = totpSecret.account ?: screenStrings.actionsSheetNoAccount,
-        icon = Icons.Default.AccountBox,
+        heading = passkey.displayName,
+        subheading = passkey.description,
+        icon = { PasskeyIcon() },
     )
     BottomSheetAction(
         icon = Icons.Default.Edit,
         text = screenStrings.actionsSheetEdit,
-        onClick = { onEditMetadata(totpSecret) },
+        onClick = { onEditDisplayName(passkey) },
     )
     BottomSheetAction(
         icon = Icons.Default.DeleteForever,
         text = screenStrings.actionsSheetDelete,
-        onClick = { onDeleteSecret(totpSecret) },
+        onClick = { onDeletePasskey(passkey) },
     )
 }
