+++ app/src/main/java/se/koditoriet/snout/ui/screens/main/secrets/sheets/SecretActionsSheet.kt
@@ -1,37 +1,37 @@
-package se.koditoriet.snout.ui.sheets
+package se.koditoriet.snout.ui.screens.main.secrets.sheets
 
 import BottomSheetAction
 import BottomSheetContextualHeader
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.AccountBox
 import androidx.compose.material.icons.filled.DeleteForever
 import androidx.compose.material.icons.filled.Edit
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.runtime.Composable
 import se.koditoriet.snout.appStrings
-import se.koditoriet.snout.ui.components.PasskeyIcon
-import se.koditoriet.snout.vault.Passkey
+import se.koditoriet.snout.vault.TotpSecret
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun PasskeyActionsSheet(
-    passkey: Passkey,
-    onEditDisplayName: (Passkey) -> Unit,
-    onDeletePasskey: (Passkey) -> Unit,
+fun SecretActionsSheet(
+    totpSecret: TotpSecret,
+    onEditMetadata: (TotpSecret) -> Unit,
+    onDeleteSecret: (TotpSecret) -> Unit,
 ) {
-    val screenStrings = appStrings.managePasskeysScreen
+    val screenStrings = appStrings.secretsScreen
     BottomSheetContextualHeader(
-        heading = passkey.displayName,
-        subheading = passkey.description,
-        icon = { PasskeyIcon() },
+        heading = totpSecret.issuer,
+        subheading = totpSecret.account ?: screenStrings.actionsSheetNoAccount,
+        icon = Icons.Default.AccountBox,
     )
     BottomSheetAction(
         icon = Icons.Default.Edit,
         text = screenStrings.actionsSheetEdit,
-        onClick = { onEditDisplayName(passkey) },
+        onClick = { onEditMetadata(totpSecret) },
     )
     BottomSheetAction(
         icon = Icons.Default.DeleteForever,
         text = screenStrings.actionsSheetDelete,
-        onClick = { onDeletePasskey(passkey) },
+        onClick = { onDeleteSecret(totpSecret) },
     )
 }
