+++ app/src/main/java/se/koditoriet/snout/ui/screens/setup/SetupScreen.kt
@@ -1,166 +1,61 @@
-@file:OptIn(ExperimentalMaterial3Api::class)
-
 package se.koditoriet.snout.ui.screens.setup
 
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.material3.Button
-import androidx.compose.material3.Card
-import androidx.compose.material3.CenterAlignedTopAppBar
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.RadioButton
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
+import androidx.activity.compose.BackHandler
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import se.koditoriet.snout.appStrings
-import se.koditoriet.snout.ui.theme.PADDING_L
-import se.koditoriet.snout.ui.theme.PADDING_XL
-import se.koditoriet.snout.ui.theme.SPACING_L
-import se.koditoriet.snout.ui.theme.SPACING_M
-import se.koditoriet.snout.ui.theme.SPACING_XXL
+import androidx.fragment.app.FragmentActivity
+import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.print.PrintHelper
+import se.koditoriet.snout.crypto.BackupSeed
+import se.koditoriet.snout.crypto.wordMap
+import se.koditoriet.snout.ui.onIOThread
+import se.koditoriet.snout.ui.components.LoadingOverlay
+import se.koditoriet.snout.viewmodel.SnoutViewModel
 
 @Composable
-fun BackupSetupScreen(
-    onEnableBackups: () -> Unit,
-    onSkipBackups: () -> Unit,
-    onRestoreBackup: () -> Unit,
-) {
-    val screenStrings = appStrings.setupScreen
-    var backupChoice by remember { mutableStateOf(BackupChoice.EnableBackups) }
+fun FragmentActivity.SetupScreen() {
+    val viewModel = viewModel<SnoutViewModel>()
+    var viewState by remember { mutableStateOf<ViewState>(ViewState.InitialSetup) }
+    var showLoadingScreen by remember { mutableStateOf(false) }
 
-    Scaffold(
-        topBar = {
-            CenterAlignedTopAppBar(
-                title = { Text(screenStrings.welcome) }
-            )
-        }
-    ) { padding ->
-        LazyColumn(
-            modifier = Modifier
-                .padding(padding)
-                .padding(PADDING_XL)
-                .fillMaxSize(),
-            verticalArrangement = Arrangement.SpaceBetween
-        ) {
-            item {
-                Column {
-                    Text(
-                        text = screenStrings.enableBackups,
-                        style = MaterialTheme.typography.headlineMedium
-                    )
-                    Spacer(Modifier.height(SPACING_L))
-                    Text(
-                        text = screenStrings.enableBackupsDescription,
-                        style = MaterialTheme.typography.bodyMedium
-                    )
-
-                    Spacer(Modifier.height(SPACING_XXL))
-
-                    BackupChoiceCard(
-                        title = screenStrings.enableBackupsCardEnable,
-                        description = screenStrings.enableBackupsCardEnableDescription,
-                        selected = backupChoice == BackupChoice.EnableBackups,
-                        onClick = { backupChoice = BackupChoice.EnableBackups }
-                    )
-
-                    Spacer(Modifier.height(SPACING_M))
-
-                    BackupChoiceCard(
-                        title = screenStrings.enableBackupsCardDisable,
-                        description = screenStrings.enableBackupsCardDisableDescription,
-                        selected = backupChoice == BackupChoice.DisableBackups,
-                        onClick = { backupChoice = BackupChoice.DisableBackups }
-                    )
-
-                    Spacer(Modifier.height(SPACING_M))
-
-                    BackupChoiceCard(
-                        title = screenStrings.enableBackupsCardImport,
-                        description = screenStrings.enableBackupsCardImportDescription,
-                        selected = backupChoice == BackupChoice.ImportAndEnableBackups,
-                        onClick = { backupChoice = BackupChoice.ImportAndEnableBackups }
-                    )
-                }
-            }
-
-            item {
-                Spacer(Modifier.height(SPACING_M))
-                Button(
-                    modifier = Modifier.fillMaxWidth(),
-                    onClick = {
-                        when (backupChoice) {
-                            BackupChoice.DisableBackups -> onSkipBackups()
-                            BackupChoice.EnableBackups -> onEnableBackups()
-                            BackupChoice.ImportAndEnableBackups -> onRestoreBackup()
-                        }
-                    }
-                ) {
-                    Text(appStrings.generic.continueOn)
-                }
-            }
+    BackHandler {
+        viewState.previousViewState?.apply {
+            viewState = this
         }
     }
-}
 
-@Composable
-private fun BackupChoiceCard(
-    title: String,
-    description: String,
-    selected: Boolean,
-    onClick: () -> Unit
-) {
-    val borderColor =
-        if (selected) {
-            MaterialTheme.colorScheme.primary
-        }  else {
-            MaterialTheme.colorScheme.outline
-        }
-
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable { onClick() },
-        border = BorderStroke(2.dp, borderColor)
-    ) {
-        Row(
-            modifier = Modifier.padding(PADDING_L),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            RadioButton(
-                selected = selected,
-                onClick = onClick
+    when (viewState) {
+        ViewState.InitialSetup -> {
+            InitialSetupScreen(
+                onEnableBackups = { viewState = ViewState.ShowBackupSeed },
+                onSkipBackups = onIOThread { viewModel.createVault(null) },
+                onRestoreBackup = { viewState = ViewState.RestoreBackup }
             )
-
-            Spacer(Modifier.width(SPACING_M))
-
-            Column {
-                Text(title, style = MaterialTheme.typography.titleMedium)
-                Text(description, style = MaterialTheme.typography.bodySmall)
-            }
+        }
+        ViewState.ShowBackupSeed -> {
+            val seed = remember { BackupSeed.generate() }
+            BackupSeedScreen(
+                backupSeed = seed,
+                onPrintQr = PrintHelper(this)::printBitmap,
+                onContinue = onIOThread {
+                    viewModel.createVault(seed)
+                    seed.wipe()
+                }
+            )
+        }
+        ViewState.RestoreBackup -> {
+            LoadingOverlay(showLoadingScreen)
+            RestoreBackupScreen(
+                seedWords = wordMap.keys,
+                onRestore = onIOThread { backupSeed, uri ->
+                    showLoadingScreen = true
+                    viewModel.restoreVaultFromBackup(backupSeed, uri)
+                    showLoadingScreen = false
+                }
+            )
         }
     }
-}
-
-private enum class BackupChoice {
-    DisableBackups,
-    EnableBackups,
-    ImportAndEnableBackups,
 }
