+++ app/src/main/java/se/koditoriet/snout/ui/screens/setup/ViewState.kt
@@ -1,0 +1,9 @@
+package se.koditoriet.snout.ui.screens.setup
+
+import se.koditoriet.snout.vault.NewTotpSecret
+
+sealed class ViewState(val previousViewState: ViewState?) {
+    object InitialSetup : ViewState(null)
+    object ShowBackupSeed : ViewState(InitialSetup)
+    object RestoreBackup : ViewState(InitialSetup)
+}
